second push

Approved-by: Jordan Ubaldo

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/User.java
@@ -6,7 +6,7 @@ import java.util.Set;
 
 public class User {
 
-   private Long id;
+   private int id;
    private String username;
    private String password;
    private boolean activated;
@@ -14,23 +14,23 @@ public class User {
 
    public User() { }
 
-   public User(Long id, String username, String password, String authorities) {
+   public User(int id, String username, String password, String authorities) {
       this.id = id;
       this.username = username;
       this.password = password;
       this.activated = true;
    }
 
-   public User(Long id, String username) {
+   public User(int id, String username) {
       this.id = id;
       this.username = username;
    }
 
-   public Long getId() {
+   public int getId() {
       return id;
    }
 
-   public void setId(Long id) {
+   public void setId(int id) {
       this.id = id;
    }
